+++ peekingduck/pipeline/utils/bbox/transforms.py
@@ -6,6 +6,21 @@ import numpy as np
 import torch
 
 
+def tlwh2xyxyn(inputs: np.ndarray, height: int, width: int) -> np.ndarray:
+    """Converts from [t, l, w, h] to [x1, y1, x2, y2] format.
+
+    (x1, y1) and (x2, y2) are coordinates of top left and bottom right
+    respectively. (t, l) is the coordinates of the top left corner, w is the
+    width, and h is the height.
+    """
+    outputs = np.empty_like(inputs)
+    outputs[:, 0] = inputs[:, 0] / width
+    outputs[:, 1] = inputs[:, 1] / height
+    outputs[:, 2] = (inputs[:, 0] + inputs[:, 2]) / width
+    outputs[:, 3] = (inputs[:, 1] + inputs[:, 3]) / height
+    return outputs
+
+
 def xywh2xyxy(inputs: torch.Tensor) -> torch.Tensor:
     """Converts from [x, y, w, h] to [x1, y1, x2, y2] format.
